read skills: rename _mcp.py -> mcp_client.py + declare it (prod-install robustness)
The vendored MCP helper failed to land on prod boxes ('No module named _mcp')
while present on dev. Two plausible causes, both fixed here so it can't recur
regardless of which it is:
- a selective installer that strips '_'-prefixed files -> renamed off the
  underscore (mcp_client.py), so nothing can mistake it for private/hidden.
- a 'fetch only declared/referenced files' install (agent- or manifest-driven)
  that missed a runtime-only import -> added an explicit Install note to each
  SKILL.md ('install the whole scripts/ dir; engine imports mcp_client.py').

Mechanical: git mv _mcp.py->mcp_client.py, 'from _mcp import'->'from mcp_client
import' in all 6 engines, README refs updated. Offline tests pass
(8/8,8/8,5/5,4/4,5/5,4/4); every mcp_client import resolves.
