+++ senpi-account-status/scripts/mcp_client.py
