+++ senpi-market-pulse/scripts/mcp_client.py
