+++ senpi-market-pulse/scripts/pulse.py
@@ -100,7 +100,7 @@ def _funding_sign(funding):
 def _get_client():
     if HERE not in sys.path:
         sys.path.insert(0, HERE)
-    from _mcp import MCPClient
+    from mcp_client import MCPClient
     return MCPClient()
