+++ senpi-portfolio/scripts/mcp_client.py
