+++ senpi-portfolio/scripts/portfolio.py
@@ -80,7 +80,7 @@ def _pct(mark, prev):
 def _get_client():
     if HERE not in sys.path:
         sys.path.insert(0, HERE)
-    from _mcp import MCPClient
+    from mcp_client import MCPClient
     return MCPClient()
