+++ senpi-smart-money/scripts/mcp_client.py
