+++ senpi-smart-money/scripts/smartmoney.py
@@ -89,7 +89,7 @@ def _traders_of(data):
 def _get_client():
     if HERE not in sys.path:
         sys.path.insert(0, HERE)
-    from _mcp import MCPClient
+    from mcp_client import MCPClient
     return MCPClient()
