+++ senpi-trader-research/scripts/mcp_client.py
